+++ examples/wrapper-types/tests/test_generated_bindings.rs
@@ -1,4 +1,4 @@
 uniffi_macros::build_foreign_language_testcases!(
-    "src/wrapper-types.udl",
+    ["src/wrapper-types.udl"],
     ["tests/bindings/test_wrapper_types.py"]
 );