+++ fixtures/coverall/tests/test_generated_bindings.rs
@@ -1,5 +1,5 @@
 uniffi_macros::build_foreign_language_testcases!(
-    "src/coverall.udl",
+    ["src/coverall.udl",],
     [
         "tests/bindings/test_coverall.py",
         "tests/bindings/test_coverall.kts",